fix(int)!: add int32 and fix int  (#353)
* storage getInfo

* fix int issue and client test

* fix helper func

* fix int types

* fix int types

* chore(deps): update server to current main

---------

+++ vela/step.go
@@ -62,7 +62,7 @@ func (svc *StepService) Add(org, repo string, build int, s *api.Step) (*api.Step
 }
 
 // Update modifies a step with the provided details.
-func (svc *StepService) Update(org, repo string, build int, s *api.Step) (*api.Step, *Response, error) {
+func (svc *StepService) Update(org, repo string, build int64, s *api.Step) (*api.Step, *Response, error) {
 	// set the API endpoint path we send the request to
 	u := fmt.Sprintf("/api/v1/repos/%s/%s/builds/%d/steps/%d", org, repo, build, s.GetNumber())
